budget_app/requirements.txt
streamlit
pandas
python-dateutil